--- ghost/audience-feedback/lib/audience-feedback.js
@@ -1,5 +1,0 @@
-module.exports = {
-    AudienceFeedbackService: require('./AudienceFeedbackService'),
-    AudienceFeedbackController: require('./AudienceFeedbackController'),
-    Feedback: require('./Feedback')
-};
