--- ghost/audience-feedback/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@tryghost/audience-feedback",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/audience-feedback",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.2.0",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/tpl": "0.1.32",
-    "bson-objectid": "2.0.4"
-  }
-}
