+++ ghost/core/package.json
@@ -70,7 +70,6 @@
     "@tryghost/announcement-bar-settings": "0.0.0",
     "@tryghost/api-framework": "0.0.0",
     "@tryghost/api-version-compatibility-service": "0.0.0",
-    "@tryghost/audience-feedback": "0.0.0",
     "@tryghost/bookshelf-plugins": "0.6.25",
     "@tryghost/captcha-service": "0.0.0",
     "@tryghost/color-utils": "0.2.2",
